feat: adding useTasks hook.
