+++ apps/tasks/frontend/src/App.tsx
@@ -3,90 +3,16 @@ import "./App.css"
 import TasksFooter from "./components/TasksFooter"
 import TasksHeader from "./components/TasksHeader"
 import TasksBody from "./components/TasksBody"
-import { useEffect, useState } from "react"
-import { Task } from "./types"
-import {
-  createTask,
-  deleteTask,
-  getAllTasks,
-  updateTask,
-} from "./repositories/fetchTaskRepository"
+import useTasks from "./hooks/useTasks"
 
 function App() {
-  const [isLoading, setIsLoading] = useState(true)
-  const [history, setHistory] = useState<Array<Array<Task>>>([])
-  const [tasks, setTasks] = useState<Array<Task>>([])
-
-  useEffect(function () {
-    getAllTasks().then((tasks) => {
-      setTasks(tasks)
-      setHistory((prev) => [...prev, tasks])
-      setIsLoading(false)
-    })
-  }, [])
-
-  const handleAddTask = async (task: Task) => {
-    try {
-      const localId = task.id
-      setTasks((prev) => [task, ...prev])
-      const newTask = await createTask(task)
-      // thor mode 🤦 => only the server know the new id
-      setTasks((prev) =>
-        prev.map((task: Task) => {
-          if (localId === task.id) {
-            return { ...task, id: newTask.id }
-          }
-
-          return task
-        })
-      )
-      setHistory((prev) => [...prev, tasks])
-    } catch (e) {
-      undo()
-    }
-  }
-
-  const handleToggleTask = async (task: Task) => {
-    try {
-      const taskId = task.id
-      setTasks((prev) =>
-        prev.map((task: Task) => {
-          if (taskId === task.id) {
-            return { ...task, completed: !task.completed }
-          }
-
-          return task
-        })
-      )
-      setHistory((prev) => [...prev, tasks])
-      await updateTask(task)
-    } catch (e) {
-      console.error(e)
-      undo()
-    }
-  }
-
-  const handleRemoveTask = async (taskId: string) => {
-    try {
-      setTasks((prev) =>
-        prev.filter((task: Task) => {
-          return taskId !== task.id
-        })
-      )
-      setHistory((prev) => [...prev, tasks])
-      await deleteTask(taskId)
-    } catch (e) {
-      console.error(e)
-      undo()
-    }
-  }
-
-  const undo = () => {
-    const lastTasks = history[history.length - 1]
-    const newHistory = history.slice(0, history.length - 1)
-    setTasks(lastTasks)
-    setHistory(newHistory)
-  }
+  const {
+    isLoading,
+    tasks,
+    handleAddTask,
+    handleToggleTask,
+    handleRemoveTask,
+  } = useTasks()
 
   if (isLoading) {
     return <div>Cargando lista de tareas...</div>
